Update per testing with Transmitter connected
I found that when the transmitter is connected to pin 3 (with the LED
also connected), it is best to leave the pin as OUTPUT and in the LOW
state (vs INPUT, no pullup).  I suspect that this cuts down on
transmitter noise as well.  Quiescent power draw (ATtiny85 +
transmitter) is about 0.5 microamp!